additional work on the directory publisher
